pd_ocr/requirements.txt
fastapi
uvicorn[standard]
textblob
pyodbc
pytest
python-decouple
easyocr
pandas
behave
sqlalchemy
minio
opencv-python
cython
aiofiles



